ascii/requirements.txt
opencv-python==4.0.0.21
Pillow==5.4.1
